book-rag/pyproject.toml
[project]
name = "book-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "fastapi[standard]>=0.124.4",
    "langchain[standard]>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-qdrant>=0.1.2",
    "openai>=1.40",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.12.0",
    "uvicorn>=0.38.0",
]
